calling replace on clientside will now either replace the whole element (like before) or optionally just replace its content
this is needed for TJuiDialog (maybe for other jqueryui controls too)
because if the element (in case of TJuiDialog the div panel) will be
replaced, jqueryui will clean up by calling destroy on dom element
removal causing the replaceWith to fail because the parent node is
already removed when replacing the original element with the new
content. to avoid this, only the inner content will be replaced which is
now controlled by an optional parameter for the clientscript replace
method.