Fix validators on some subclasses; Fix #683 (#684)
Client side validators needs special handling to extract values from some special controls (eg. lists). After the "big namespace update" this code worked in our tests, but failed to correctly detect subclasses of these special controls.
This PR reworks the code in order to fix the issue and clean up  a bit the mess.

Also, an additional change removes the usage of PromptText as a fallback for PromptValue if not specified. This was introduced for browser support, but it looks like modern browsers can handle empty values just fine, while this behavior tricks our validators.