#815 Checks if the shell ENV from crontab to trigger "only tasks executing in current minute" logic
Basically this is required because we know each minute cron is running.  If a tasks takes longer than the current executing minute, the further tasks should not be executed due to being picked up by the next crontab prado cron.

This preserves the CLI functionality that when PRADO cron is manually run from the CLI, all pending tasks are completed regardless of when its run or how long it takes.