#834 TBehaviorsModule Attaching behaviors to behaviors
behavior names are unique to the owner object.   But To reference behaviors to attach second layer behaviors, the behavior names must be unique across the configured TBehaviorsModule behaviors.  Multiple TBehaviorsModules (eg. a second one in TPageService for page specific behaviors) can refer to the behaviors of the first TBehaviorsModule.

With unit Tests