#843 changed PermissionsBehaviors::getManager to getPermissionsManager for specificity (#883)
* #843 changed Permissions::getManager to getPermissionsManager for specificity

basically, as behaviors, these shouldn't hog the "getManager" name from other behaviors, and it needed to be more specific on its own.

* #844 new TPermissionsBehavior dyEvent for logging permissions failures

instead of each module/action logging its own permissions failures, raise this dynamic event and the failure is logged in the application and on the shell, as appropriate.  Shell users are displayed the failed permission only when the app is in Debug mode.