#889 TList/TMap variable types and minor updates (#890)
* #889 TList/TMap variable types and minor updates

TList and TMap variables and some methods are typed.  subclass methods that could override an input or output are not typed.

The storage is made protected rather than private because some subclasses may need direct access to the storage _d.  TWeakList needs to scrub _d in place without accessor functions.

A few functions were move around to make the files and method locations more symmetric to each other.

_getZappableSleepProps is added to TList, TPriorityList, TPriorityMap, and TWeakCallableCollections

TPriorityList uses its parent _d.  Smaller code can be made for insertAtIndexInPriority as well.  offsetGet is redundant

Regression on existing TList and TMap subclasses are updated

TPriorityListTest unit tests are corrected (not that it did something erroneous)

* #889 Reverted removal of TPropertyValue

* #889 two other edits.

* #890 corrected setReadOnly to public

* #890 slightly faster TPriorityList::removeAt based upon how TList works

* #889 Updated Unit Tests on TList::ReadOnly for text string and should be protected

TList::Read only has been historically protected so a list cannot be made mutable on the fly.  A copy can be made and made changable.