#885 TWeakList scrubbing of invalid WeakReferences, by default depends on ! Read Only
By default, a mutable TWeakList should scrub all its items when a WeakReference is lost, but when ReadOnly that will result in changes to the list.

Thus, When read only, by default, the list is not scrubbed of invalid WeakReference. and when mutable, by default, the list is scrubbed.

The default behavior can be overridden and whether to scrub can be directly specified.