TPriorityList and TPriorityMap updates for less variable copying and better organization (#899)
* TWeakList added to classes.php,  better documentation on unit tests

* #897 Updated TPriorityList and TPriorityMap, TPriorityCollection abstraction

This adds IPriorityCollection interface to TPriorityList and TPriorityMap.
 TPriorityMap can copyFrom and mergeWith IPriorityCollection but TPriorityList can only do so with itself.

The common class properties and methods are abstracted into TPriorityCollectionTrait.

After a thorough review of TPriorityList and TPriorityMap There were some speed boosts i found.  Some loops replicated data that didn't need to be.  So this will speed things up a bit and reduce PHP memory allocations/deallocations/copy.

This also adds some cross compatibility (copy and merge) between TPriorityList and TPriorityMap.

* #897 Flattened Priority Cache is made more efficient

There is less memory/copying going on around the flattened priority cache.

* #897 updated TPriorityList Unit Test

* #897 TPriorityList::itemAtIndexInPriority mimics TList::itemAt for Out-Of-Range

Added an error message.
Updated unit test for the error when out of range.

* #897 Slightly better error message on TPriorityList out of range

* #897 Bug in TPriorityList::insertAt(0) when no items in the list.

* #897 Better Bug fix for TPriorityList::insertAt

* #897 TList::priorityOf/priorityAt efficiency boost

* #897 TPriorityList::priorityAt better fix

* #897 TPriorityList::insertAtIndexInPriority Better error checking

With unit tests.

* #897 Better error checking for TPriorityList::insertAtIndexInPriority

and better documentation for priorityOf

* #897 TPriorityList::priorityAt is self containing the Count

The getCount should be _c so as to not trigger the children into Count scrubbing.  This way there is a non-scrubbing TPriorityList and a scrubbing sub-class version.  Without this change, there would only be a scrubbing version of this priorityAt

* #897 TPriorityCollectionTrait Above/Below speed bump

This does one merge rather than many.

* TWeakCallableCollection @ since  on added functions.

Upgraded php-cs-fixer and got these updates.