#827 TTemplateManager accepts template class and culture (#902)
* #827 TTemplateManager accepts template class and culture

With the option of setting a new default template class upon instancing of TTemplateManager.

This change allows a template to be formed upon a designated culture rather than just the web user's culture settings.  this is important for when sending emails to someone with different culture settings.  The template must be formed around the email receiver and not the web user.

* #827 Updated TTemplateManager to check for ITemplate rather than TTemplate

ITemplate adds getIncludedFiles method.

The templates derived from ITemplate needed new getIncludedFiles.

* #827 rm whitespace from update, PHP-CS-Fixer update