#904 TRational unit classes Prado\Util\Math
A Rational is found in EXIF and is common in its GPS "directory".  This parses and generates rationals and unsigned rationals into various formats, float, string (eg. `13/8`) and array (eg. `[13, 8]`).

The star method of this class is the Continued Fraction computation to generate the integer numerator and integer denominator from a float, to a specific tolerance.

Updated TBitHelper with specific max, min, unsigned max, and mask for 32 bit and 64 bit integers, when the values are needed and not just the system values.